Update the Wiki links
The page isn't that big.  The on-page reference links for the wiki might be confusing to the users as it doesn't look like clicking the link is taking you anywhere (and only one link takes you to the Wiki - the smallest and least notible one).

Change all the wiki links to point to the repo wiki directly.  This is probably more intuitive.

The other page reference links are probably ok.  They are just point the users to the section on the page that describes the features contained within the repo.